+++ src/main/java/wang/kingweb/community/dto/PaginationDTO.java
@@ -11,29 +11,29 @@ public class PaginationDTO {
     private boolean hasNextPage; //是否有下一页
     private boolean hasFirstPage;   //是否显示首页
     private boolean hasLastPage;    //是否显示尾页
-    private int offset; //mysql分页语法中的偏移量
-    private int page;   //页码
-    private int size;   //每页显示数据数
-    private int countPage;  //总页数
-    private List<Integer> pageList = new ArrayList<>(); //分页列表中的页码
+    private Long offset; //mysql分页语法中的偏移量
+    private Long page;   //页码
+    private Long size;   //每页显示数据数
+    private Long countPage;  //总页数
+    private List<Long> pageList = new ArrayList<>(); //分页列表中的页码
 
     //具体计算分页数据类，并设置值
-    public void setPagination(int page, int size, int totalCount) {
+    public void setPagination(Long page, Long size, Long totalCount) {
         this.page = page;
 
         //计算总页数
-        if(totalCount%size==0){
+        if(totalCount%size==0L){
             countPage = totalCount/size;
         }else{
             countPage = totalCount/size +1;
         }
-        if(totalCount==0){
-            countPage = 0;
+        if(totalCount==0L){
+            countPage = 0L;
         }
 
         //限制页码范围，防止用户输入导致越界
-        if(page<1||countPage==0){
-            page = 1;
+        if(page<1L||countPage==0L){
+            page = 1L;
         }
         else if (page>countPage){
             page = countPage;
