+++ src/main/java/wang/kingweb/community/service/PaginationService.java
@@ -6,7 +6,7 @@ import wang.kingweb.community.dto.PaginationDTO;
 @Service
 public class PaginationService {
 
-    public PaginationDTO getPageInfo(int page, int size, int totalCount) {
+    public PaginationDTO getPageInfo(Long page, Long size, Long totalCount) {
         PaginationDTO paginationDTO = new PaginationDTO();
         paginationDTO.setPagination(page,size,totalCount);
         return paginationDTO;
